Changed error code for parsing error when doing Script execution Removed commented code in JsonCommandRenderer Changed so NetworkProcessor.Dichotomize() uses non-validating _addEdge instead of AddEdge (thus similar to Symmetrize etc) Removed writing 'Exiting...' to console when exiting Threadle - just messes up JSON etc Added IsModified=false in Nodeset constructur (that comes with a set of nodes without attributes) Fixed so that WordWrap() in TextCommand..Renderer keeps initial intendation Changed weights to long rather than int in GetRandomEdgeWeightedTwoMode

+++ Threadle.Core/Analysis/Functions.cs
@@ -480,14 +480,14 @@ namespace Threadle.Core.Analysis
         {
             string[] allNames = layerTwoMode.GetAllHyperedgeNames(0, (int)Math.Min(layerTwoMode.NbrHyperedges, int.MaxValue));
             List<string> validNames = [];
-            List<int> weights = [];
+            List<long> weights = [];
             long totalWeight = 0;
             foreach (string name in allNames)
             {
                 int k = layerTwoMode.GetHyperedgeNodeIds(name).Length;
                 if (k >= 2)
                 {
-                    int weight = k * (k - 1) / 2;
+                    long weight = k * (k - 1) / 2;
                     validNames.Add(name);
                     weights.Add(weight);
                     totalWeight += weight;
